# backend/openapi.generated.yaml
openapi: 3.1.0
info:
  title: CloudPulse AI API
  description: 'Serverless cloud governance, cost, and compliance platform for AWS.


    Every endpoint except `/health` requires a Cognito-issued bearer token. The caller''s

    role is derived from directory group membership on every request and is never
    stored by

    the platform (FR-031a). A token whose group claim contains zero or more than one
    mapped

    group is refused (FR-032a).

    '
  version: 0.1.0
paths:
...
  /resources/{resource_id}/owner:
    get:
      tags:
      - ownership
      summary: Retrieve a resource's attributed owner
      description: 'FR-020-FR-028. Any role may view (FR-030). 200 with a null `ownerEmail`/

        `confidence` (not 404) when the resource exists but is queued

        unattributed (FR-022) -- a resource genuinely not found still 404s.'
      operationId: getResourceOwner
      parameters:
      - name: resource_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Resource Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceOwnership'
        '401':
          description: No valid identity presented.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Valid identity, but the caller's role does not permit the action
            -- or the caller maps to zero or multiple directory groups and therefore
            has no resolvable role (FR-032a, FR-034).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Target record does not exist, or the caller is not entitled
            to know that it does.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    ErrorBody:
      properties:
        code:
          type: string
          title: Code
          description: Stable and machine-readable. SCREAMING_SNAKE_CASE.
        message:
          type: string
          title: Message
          description: Human-readable and safe to display. Never reveals whether a
            given record exists to a caller not entitled to know (FR-035).
        correlationId:
          type: string
          title: Correlationid
          description: Matches the id in the structured logs for this request (FR-044).
        details:
          anyOf:
          - items:
              $ref: '#/components/schemas/ErrorDetail'
            type: array
          - type: 'null'
          title: Details
          description: Field-level detail. Present for validation failures only.
      type: object
      required:
      - code
      - message
      - correlationId
      title: ErrorBody
    ErrorDetail:
      properties:
        field:
          type: string
          title: Field
        issue:
          type: string
          title: Issue
      type: object
      required:
      - field
      - issue
      title: ErrorDetail
      description: One field-level validation problem.
    ErrorEnvelope:
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      type: object
      required:
      - error
      title: ErrorEnvelope
      description: 'The uniform error envelope (FR-043).


        Declared as a model, not just built as a dict, so it appears in the **generated**

        OpenAPI document -- which FR-048 makes the binding contract. An envelope that
        only

        exists at runtime would be invisible to the generated client, leaving every

        consumer to hand-roll the error type.'
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OwnershipEvidence:
      properties:
        kind:
          type: string
          title: Kind
        cloudtrailEventId:
          anyOf:
          - type: string
          - type: 'null'
          title: Cloudtraileventid
        principal:
          anyOf:
          - type: string
          - type: 'null'
          title: Principal
        eventTime:
          anyOf:
          - type: string
          - type: 'null'
          title: Eventtime
      type: object
      required:
      - kind
      title: OwnershipEvidence
    ResourceOwnership:
      properties:
        resourceId:
          type: string
          title: Resourceid
        ownerEmail:
          anyOf:
          - type: string
          - type: 'null'
          title: Owneremail
        confidence:
          anyOf:
          - type: string
          - type: 'null'
          title: Confidence
        evidence:
          anyOf:
          - $ref: '#/components/schemas/OwnershipEvidence'
          - type: 'null'
        attributedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Attributedat
      type: object
      required:
      - resourceId
      title: ResourceOwnership
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    cognitoJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito-issued access token. Validated by the API Gateway JWT authorizer,
        then re-validated for role cardinality by the application (research.md R-004).
security:
- cognitoJwt: []